requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
python-decouple==3.8
rest_framework_simplejwt==0.0.2
